feat: #44 ミドルウェアの定義を分離し、tasks.tsで設定するよう変更
