+++ backend/src/presentations/middlewares/cors.ts
@@ -1,0 +1,11 @@
+import { cors } from 'hono/cors'
+import { createMiddleware } from 'hono/factory'
+
+export const corsMiddleware = createMiddleware<{ Bindings: Env }>(
+  async (c, next) => {
+    const corsMiddleware = cors({
+      origin: c.env.CORS_ALLOW_ORIGINS,
+      credentials: true,
+    })
+    return corsMiddleware(c, next)
+  })
